refactor(normalization-core): promote shared result type
