+++ packages/agent-core/src/harness/types.ts
@@ -1,4 +1,5 @@
 // Agent Core type module defines shared TypeScript contracts.
+import type { Result } from "@openclaw/normalization-core/result";
 import type {
   ImageContent,
   Model,
@@ -11,18 +12,8 @@ import type { AgentEvent, AgentMessage, AgentTool, QueueMode, ThinkingLevel } fr
 import type { AgentCoreCompletionRuntimeDeps, AgentCoreRuntimeDeps } from "../runtime-deps.js";
 import type { Session } from "./session/session.js";
 
-/** Result of a fallible operation. Expected failures are returned as `ok: false` instead of thrown. */
-export type Result<TValue, TError> = { ok: true; value: TValue } | { ok: false; error: TError };
-
-/** Create a successful {@link Result}. */
-export function ok<TValue, TError>(value: TValue): Result<TValue, TError> {
-  return { ok: true, value };
-}
-
-/** Create a failed {@link Result}. */
-export function err<TValue, TError>(error: TError): Result<TValue, TError> {
-  return { ok: false, error };
-}
+export { err, ok } from "@openclaw/normalization-core/result";
+export type { Result } from "@openclaw/normalization-core/result";
 
 /** Normalize unknown thrown values into Error instances before using them as typed error causes. */
 export function toError(error: unknown): Error {
