+++ tsconfig.json
@@ -145,6 +145,9 @@
       "@openclaw/normalization-core/record-coerce": [
         "./packages/normalization-core/src/record-coerce.ts"
       ],
+      "@openclaw/normalization-core/result": [
+        "./packages/normalization-core/src/result.ts"
+      ],
       "@openclaw/normalization-core/string-coerce": [
         "./packages/normalization-core/src/string-coerce.ts"
       ],
